Extract reports form into separate component
