+++ site/gatsby-site/src/components/incidents/IncidentForm.js
@@ -3,7 +3,7 @@ import { Form as FormikForm, useFormikContext } from 'formik';
 import * as Yup from 'yup';
 import SemanticallyRelatedIncidents from '../SemanticallyRelatedIncidents';
 import RelatedIncidentsArea from '../RelatedIncidentsArea';
-import { gql, useQuery, useMutation } from '@apollo/client';
+import { gql, useQuery } from '@apollo/client';
 import FieldContainer from 'components/forms/SubmissionWizard/FieldContainer';
 import TextInputGroup from 'components/forms/TextInputGroup';
 import { useTranslation } from 'react-i18next';
@@ -11,14 +11,8 @@ import TagsInputGroup from 'components/forms/TagsInputGroup';
 import Label from 'components/forms/Label';
 import UsersField from 'components/users/UsersField';
 import IncidentsField from './IncidentsField';
-import { Spinner, Button, Badge } from 'flowbite-react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faTag, faEdit, faExternalLinkAlt } from '@fortawesome/free-solid-svg-icons';
-import { UPDATE_REPORT } from '../../graphql/reports';
-import { Typeahead } from 'react-bootstrap-typeahead';
-import useToastContext, { SEVERITY } from '../../hooks/useToast';
-import { isCompleteReport } from '../../utils/variants';
-import { getUnixTime } from 'date-fns';
+import LinkedReportsForm from './LinkedReportsForm';
+import { Spinner } from 'flowbite-react';
 
 const relatedIncidentIdsQuery = gql`
   query IncidentWithReports($filter: IncidentFilterType) {
@@ -76,90 +70,6 @@ function IncidentForm({ entityNames = [] }) {
   const [editorSimilarIncidents, setEditorSimilarIncidents] = useState([]);
 
   const [similarIncidentsById, setSimilarIncidentsById] = useState([]);
-
-  const addToast = useToastContext();
-
-  const [updateReport] = useMutation(UPDATE_REPORT);
-
-  const [editingReportTags, setEditingReportTags] = useState(null);
-
-  const [reportTags, setReportTags] = useState({});
-
-  const [originalTags, setOriginalTags] = useState({});
-
-  const [allTags, setAllTags] = useState([]);
-
-  useEffect(() => {
-    if (values.reports) {
-      const tags = new Set();
-
-      values.reports.forEach((report) => {
-        if (report.tags) {
-          report.tags.forEach((tag) => tags.add(tag));
-        }
-      });
-
-      setAllTags(Array.from(tags));
-
-      const tagsMap = {};
-
-      values.reports.forEach((report) => {
-        tagsMap[report.report_number] = Array.isArray(report.tags) ? [...report.tags] : [];
-      });
-      setReportTags(tagsMap);
-    }
-  }, [values.reports]);
-
-  const handleUpdateReportTags = async (reportNumber, newTags) => {
-    console.log('Updating tags for report:', reportNumber, newTags);
-    try {
-      await updateReport({
-        variables: {
-          filter: {
-            report_number: { EQ: reportNumber },
-          },
-          update: {
-            set: {
-              tags: newTags,
-              epoch_date_modified: getUnixTime(new Date()),
-            },
-          },
-        },
-      });
-
-      setReportTags({
-        ...reportTags,
-        [reportNumber]: newTags,
-      });
-
-      if (values.reports) {
-        const updatedReports = values.reports.map((report) => {
-          if (report.report_number === reportNumber) {
-            return {
-              ...report,
-              tags: newTags,
-            };
-          }
-          return report;
-        });
-
-        values.reports = updatedReports;
-      }
-
-      setEditingReportTags(null);
-
-      addToast({
-        message: t('Tags updated successfully'),
-        severity: SEVERITY.success,
-      });
-    } catch (error) {
-      addToast({
-        message: t('Error updating tags: {{message}}', { message: error.message }),
-        severity: SEVERITY.danger,
-        error,
-      });
-    }
-  };
 
   useEffect(() => {
     let similarIncidentsById = [];
@@ -354,145 +264,8 @@ function IncidentForm({ entityNames = [] }) {
           />
         </FieldContainer>
 
-        {values.reports && values.reports.length > 0 && Object.keys(reportTags).length > 0 && (
-          <FieldContainer>
-            <Label label={t('Linked Reports')} popover="linkedReports" />
-            <div className="mt-2 space-y-4">
-              {values.reports
-                .filter((r) => isCompleteReport(r))
-                .map((report) => {
-                  return (
-                    <div key={report.report_number} className="border rounded-lg p-4 bg-gray-50">
-                      <div className="flex justify-between items-start mb-2">
-                        <div>
-                          <h4 className="text-lg font-medium">
-                            <a
-                              href={report.url}
-                              target="_blank"
-                              rel="noopener noreferrer"
-                              className="hover:text-blue-600 flex items-center"
-                            >
-                              {report.title}
-                              <FontAwesomeIcon icon={faExternalLinkAlt} className="ml-2 text-sm" />
-                            </a>
-                          </h4>
-                          <p className="text-sm text-gray-500">Report #{report.report_number}</p>
-                        </div>
-                        <a
-                          href={`/cite/edit?report_number=${report.report_number}`}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-blue-600 hover:text-blue-800"
-                        >
-                          <FontAwesomeIcon icon={faEdit} className="mr-1" />
-                          {t('Edit Report')}
-                        </a>
-                      </div>
-
-                      <div className="mt-3">
-                        <div className="flex items-center mb-2">
-                          <FontAwesomeIcon
-                            fixedWidth
-                            icon={faTag}
-                            title={t('Tags')}
-                            className="mr-2"
-                          />
-                          <span className="font-medium">{t('Tags')}:</span>
-                        </div>
-
-                        {editingReportTags === report.report_number ? (
-                          <div className="mt-2">
-                            <div>
-                              <Typeahead
-                                key={`report-tags-typeahead-${report.report_number}`}
-                                id={`report-tags-${report.report_number}`}
-                                className="Typeahead"
-                                allowNew
-                                multiple
-                                onChange={(value) => {
-                                  console.log('Typeahead onChange:', value);
-                                  setReportTags({
-                                    ...reportTags,
-                                    [report.report_number]: value.map((v) =>
-                                      v.label ? v.label : v
-                                    ),
-                                  });
-                                }}
-                                options={allTags}
-                                selected={reportTags[report.report_number] || []}
-                                placeholder={t('Choose or add tags...')}
-                              />
-                            </div>
-                            <div className="flex mt-2 space-x-2">
-                              <Button
-                                size="xs"
-                                onClick={() =>
-                                  handleUpdateReportTags(
-                                    report.report_number,
-                                    reportTags[report.report_number]
-                                  )
-                                }
-                              >
-                                {t('Save')}
-                              </Button>
-                              <Button
-                                size="xs"
-                                color="light"
-                                onClick={() => {
-                                  if (originalTags[report.report_number]) {
-                                    setReportTags({
-                                      ...reportTags,
-                                      [report.report_number]: [
-                                        ...originalTags[report.report_number],
-                                      ],
-                                    });
-                                  }
-                                  setEditingReportTags(null);
-                                }}
-                              >
-                                {t('Cancel')}
-                              </Button>
-                            </div>
-                          </div>
-                        ) : (
-                          <div className="flex flex-wrap gap-2 mt-1 items-center">
-                            {reportTags[report.report_number] &&
-                            reportTags[report.report_number].length > 0 ? (
-                              reportTags[report.report_number].map((tag, index) => (
-                                <Badge key={index} color="info" className="text-xs">
-                                  {tag}
-                                </Badge>
-                              ))
-                            ) : (
-                              <span className="text-gray-500 text-sm italic">{t('No tags')}</span>
-                            )}
-                            {editingReportTags !== report.report_number && (
-                              <Button
-                                size="xs"
-                                color="light"
-                                className="ml-2"
-                                onClick={() => {
-                                  setOriginalTags({
-                                    ...originalTags,
-                                    [report.report_number]: [
-                                      ...(reportTags[report.report_number] || []),
-                                    ],
-                                  });
-                                  setEditingReportTags(report.report_number);
-                                }}
-                              >
-                                <FontAwesomeIcon icon={faEdit} className="mr-1" />
-                                {t('Edit')}
-                              </Button>
-                            )}
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  );
-                })}
-            </div>
-          </FieldContainer>
+        {values.reports && values.reports.length > 0 && (
+          <LinkedReportsForm reports={values.reports} />
         )}
 
         <div id="similar-incidents">
